ECS refactor

+++ jams/space/Scene.h
@@ -8,7 +8,7 @@
 using namespace mud;
 using namespace toy;
 
-void fill_fleet(VisuFleet& visu, const std::map<ShipSchema*, size_t>& ships);
+void fill_fleet(VisuFleet& visu, const std::map<ShipSchema*, uint32_t>& ships);
 
 void paint_fleet_ships(Gnode& parent, VisuFleet& visu, float radius = 0.4f, float ship_scale = 0.05f);
